# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OTUS Highload Architect",
    "version": "1.2.0"
  },
  "paths": {
...
    "/post/list/{user_id}": {
      "get": {
        "operationId": "getUserPosts",
        "description": "Посты конкретного пользователя, от новых к старым",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "schema": {
              "$ref": "#/components/schemas/UserId"
            },
            "required": true,
            "in": "path"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получены посты пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostFeed"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
    "/dialog/list": {
      "get": {
        "operationId": "getDialogList",
        "description": "Список диалогов текущего пользователя. Возвращает собеседника и идентификатор диалога без превью последнего сообщения.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получен список диалогов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DialogSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
...
    "/dialog/{user_id}/list": {
      "get": {
        "operationId": "getDialogMessages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "schema": {
              "$ref": "#/components/schemas/UserId"
            },
            "required": true,
            "in": "path"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Диалог между двумя пользователями",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DialogMessageList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "number",
          "minimum": 0,
          "description": "Оффсет с которого начинать выдачу",
          "example": 100,
          "default": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "number",
          "minimum": 1,
          "description": "Лимит, ограничивающий кол-во возвращенных сущностей",
          "example": 10,
          "default": 10
        }
      }
    },
    "responses": {
      "400": {
        "description": "Невалидные данные ввода"
      },
      "401": {
        "description": "Неавторизованный доступ"
      },
      "5xx": {
        "description": "Ошибка сервера",
        "headers": {
          "Retry-After": {
            "description": "Время, через которое еще раз нужно сделать запрос",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Описание ошибки"
                },
                "request_id": {
                  "type": "string",
                  "description": "Идентификатор запроса. Предназначен для более быстрого поиска проблем."
                },
                "code": {
                  "type": "integer",
                  "description": "Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем."
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserId": {
        "type": "string",
        "description": "Идентификатор пользователя"
      },
...
      "PostId": {
        "type": "string",
        "description": "Идентификатор поста",
        "example": "1d535fd6-7521-4cb1-aa6d-031be7123c4d"
      },
      "PostText": {
        "type": "string",
        "description": "Текст поста",
        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices eros in cursus turpis massa."
      },
      "DialogMessageText": {
        "type": "string",
        "description": "Текст сообщения",
        "example": "Привет, как дела?"
      },
      "DialogMessage": {
        "type": "object",
        "required": ["from", "to", "text"],
        "properties": {
          "from": {
            "$ref": "#/components/schemas/UserId"
          },
          "to": {
            "$ref": "#/components/schemas/UserId"
          },
          "dialog": {
            "type": "string",
            "description": "Идентификатор диалога"
          },
          "text": {
            "$ref": "#/components/schemas/DialogMessageText"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Post": {
        "type": "object",
        "description": "Пост пользователя",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/PostId"
          },
          "text": {
            "$ref": "#/components/schemas/PostText"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserId"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PostFeed": {
        "type": "object",
        "description": "Страница ленты постов",
        "required": ["posts", "limit", "offset"],
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "DialogMessageList": {
        "type": "object",
        "description": "Страница сообщений диалога",
        "required": ["messages", "limit", "offset"],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DialogMessage"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "DialogSummary": {
        "type": "object",
        "description": "Диалог текущего пользователя с другим пользователем",
        "required": ["dialog_id", "user_id"],
        "properties": {
          "dialog_id": {
            "type": "string",
            "description": "Идентификатор диалога"
          },
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Авторизация по токену, который был получен в методе /login"
      }
    }
  }
}